Simulate data variables with known distributions

Simulate data variables with known distributions

sim_distr(n, k, distr, ...)

Arguments

n
Number of observations.

k
Number of variables.

distr
A vector of distributions to be used in each variable. It could be "normal", "exponential", "poisson", "weibull", "chisquare", "gamma" or "uniform". In case to use the same distribution for all variables, set it as distr = "distr".

...
Arguments specific to each distribution. See rnorm, rexp, rpois, rweibull, rchisq, rgamma, runif to know these arguments.

Value

sim_distr gives a data frame with the observations as rows and the variables as columns

See also